src/processor/requirements.txt
eli5
pyyaml
kafka-python
redis